pyproject.toml
[tool.poetry]
name = "funlocker"
version = "0.1.0"
description = ""
authors = ["Sangjoon Han <jphan32@me.com>"]
license = "MIT License"
readme = "README.md"
repository = "https://github.com/jphan32/FUnLocker"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pillow = "^10.0.1"
opencv-python = "^4.8.1.78"
opencv-contrib-python = "^4.8.1.78"
scipy = "^1.11.3"
tflite-runtime = { url = "https://github.com/google-coral/pycoral/releases/download/v2.0.0/tflite_runtime-2.5.0.post1-cp39-cp39-linux_aarch64.whl" }
pycoral = { url = "https://github.com/google-coral/pycoral/releases/download/v2.0.0/pycoral-2.0.0-cp39-cp39-linux_aarch64.whl" }
rpi-gpio = "^0.7.1"
keyboard = "^0.13.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
